+++ apps/ced-portal-fe/src/utils/index.ts
@@ -1,2 +1,4 @@
 export * from './urlValidator';
 export * from './compactQueryParams';
+export * from './getBenefitsDetailData';
+export * from './getLocalizedMetadataDetails';